requirements.txt
Django==4.1.9
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
pymongo==3.12.3
djongo== 1.3.6